Basic rate testing, multiple time periods, pileup
added miniaod requirements to get at primary vertex information among a few other things. Most of this is still a lot of code duplication and could be rendered much simpler with some tools for building inputs from JSON (which I am including) but do not have the full facilities for yet. Also added the current plotting scripts. Which are also terrible.